pyproject.toml
[project]
name = "buddyimagine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aioboto3>=15.5.0",
    "aiogram>=3.22.0",
    "aiohttp>=3.12.15",
    "convex>=0.7.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pillow>=11.0.0",
    "pydantic-ai>=1.21.0",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "rapidfuzz>=3.14.3",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
]
